requirements/base.txt
python-dotenv
flask
gunicorn==21.2.0